+++ electron/build-electron-release.sh
@@ -57,12 +57,12 @@ pushd "$FINAL_OUTPUT" >/dev/null
 if [ -e "mac" ]; then
     pushd "mac" >/dev/null
     if [ -e "${PDT_NAME}-${APP_VERSION}.dmg" ]; then
-        mv "${PDT_NAME}-${APP_VERSION}.dmg" "../${PKG_NAME}-${APP_VERSION}-gui-osx-x64.dmg"
+        mv "${PDT_NAME}-${APP_VERSION}.dmg" "../${PKG_NAME}-${APP_VERSION}-gui-electron-osx-x64.dmg"
     elif [ -e "${PDT_NAME}.app" ]; then
         if [[ "$OSTYPE" == "darwin"* ]]; then
-            tar czf "../${PKG_NAME}-${APP_VERSION}-gui-osx-x64.zip" "${PDT_NAME}.app"
+            tar czf "../${PKG_NAME}-${APP_VERSION}-gui-electron-osx-x64.zip" "${PDT_NAME}.app"
         elif [[ "$OSTYPE" == "linux"* ]]; then
-            tar czf "../${PKG_NAME}-${APP_VERSION}-gui-osx-x64.zip" --owner=0 --group=0 "${PDT_NAME}.app"
+            tar czf "../${PKG_NAME}-${APP_VERSION}-gui-electron-osx-x64.zip" --owner=0 --group=0 "${PDT_NAME}.app"
         fi
     fi
     popd >/dev/null
@@ -70,7 +70,7 @@ if [ -e "mac" ]; then
 fi
 
 IMG="${PKG_NAME}-${APP_VERSION}-x86_64.AppImage"
-DEST_IMG="${PKG_NAME}-${APP_VERSION}-gui-linux-x64.AppImage"
+DEST_IMG="${PKG_NAME}-${APP_VERSION}-gui-electron-linux-x64.AppImage"
 if [ -e $IMG ]; then
     mv "$IMG" "$DEST_IMG"
     chmod +x "$DEST_IMG"
@@ -79,7 +79,7 @@ fi
 EXE="${PDT_NAME} Setup ${APP_VERSION}.exe"
 if [ -e "$EXE" ]; then
     if [ ! -z $WIN32_ELN ] && [ ! -z $WIN64_ELN ]; then
-        mv "$EXE" "${PKG_NAME}-${APP_VERSION}-gui-win-setup.exe"
+        mv "$EXE" "${PKG_NAME}-${APP_VERSION}-gui-electron-win-setup.exe"
     elif [ ! -z $WIN32_ELN ]; then
         mv "$EXE" "${WIN32_ELN}.exe"
     elif [ ! -z $WIN64_ELN ]; then
@@ -90,7 +90,7 @@ fi
 # rename dmg file name
 DMG="${PKG_NAME}-${APP_VERSION}.dmg"
 if [ -e "$DMG" ]; then
-    mv "$DMG" "${PKG_NAME}-${APP_VERSION}-gui-osx.dmg"
+    mv "$DMG" "${PKG_NAME}-${APP_VERSION}-gui-electron-osx.dmg"
 fi
 
 # delete app zip file